habitat_baselines/rl/requirements.txt
lmdb==0.98
moviepy>=1.0.1
msgpack_numpy
torch==1.2.0
torchvision==0.2.2.post3
# full tensorflow required for tensorboard video support
tensorflow==1.13.1
tb-nightly
